+++ database/spring-cloud-stream-binder-oracle-txeventq/src/test/java/com/oracle/database/spring/cloud/stream/binder/Util.java
@@ -29,15 +29,15 @@ import java.sql.SQLException;
 import java.time.Duration;
 import java.util.UUID;
 
+import com.oracle.database.spring.testcontainers.OracleContainer;
 import oracle.ucp.jdbc.PoolDataSource;
 
-import org.testcontainers.oracle.OracleContainer;
 import org.testcontainers.utility.MountableFile;
 
 public class Util {
 	@SuppressWarnings("resource")
     public static OracleContainer oracleContainer() {
-        return new OracleContainer("gvenzl/oracle-free:23.26.2-slim-faststart")
+        return new OracleContainer()
                 .withStartupTimeout(Duration.ofMinutes(2)) // Needed for M1 Mac
                 .withUsername("testuser")
                 .withPassword("testpwd");