+++ database/starters/oracle-spring-boot-starter-samples/oracle-spring-boot-sample-spatial/src/test/java/com/oracle/database/spring/spatial/SpatialSampleApplicationTest.java
@@ -6,6 +6,7 @@ import java.net.URI;
 import java.time.Duration;
 import java.util.List;
 
+import com.oracle.database.spring.testcontainers.OracleContainer;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
@@ -17,7 +18,6 @@ import org.testcontainers.junit.jupiter.Testcontainers;
 import org.springframework.web.client.RestClient;
 import org.springframework.web.client.RestClientResponseException;
 import org.springframework.web.util.UriComponentsBuilder;
-import org.testcontainers.oracle.OracleContainer;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -27,7 +27,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 public class SpatialSampleApplicationTest {
     @Container
     @ServiceConnection
-    static OracleContainer oracleContainer = new OracleContainer("gvenzl/oracle-free:23.26.2-full-faststart")
+    static OracleContainer oracleContainer = new OracleContainer()
             .withStartupTimeout(Duration.ofMinutes(2))
             .withUsername("testuser")
             .withPassword("testpwd");